+++ Admin/ProductWithVariantsContentNavigationProvider.php
@@ -14,40 +14,29 @@ namespace Sulu\Bundle\ProductBundle\Admin;
 use Sulu\Bundle\AdminBundle\Navigation\ContentNavigationItem;
 use Sulu\Bundle\AdminBundle\Navigation\ContentNavigationProviderInterface;
 
-class ProductWithVariantsContentNavigationProvider implements ContentNavigationProviderInterface
+class ProductWithVariantsContentNavigationProvider extends ProductContentNavigationProvider implements ContentNavigationProviderInterface
 {
     /**
      * {@inheritdoc}
      */
     public function getNavigationItems(array $options = [])
     {
-        $details = new ContentNavigationItem('content-navigation.product.general');
-        $details->setId('details');
-        $details->setAction('details');
-        $details->setPosition(10);
-        $details->setComponent('products/components/detail-form@suluproduct');
-        $details->setResetStore(false);
+        // Product with variants has all the navigation items a normal product has.
+        $navigationItems = parent::getNavigationItems($options);
 
         $variants = new ContentNavigationItem('content-navigation.product.variants');
         $variants->setId('variants');
         $variants->setAction('variants');
-        $variants->setPosition(20);
+        $variants->setPosition(19);
         $variants->setComponent('products/components/variants-list@suluproduct');
         $variants->setDisplay(['edit']);
         $variants->setResetStore(false);
 
-        $attributes = new ContentNavigationItem('content-navigation.product.attributes');
-        $attributes->setId('attributes');
-        $attributes->setAction('attributes');
-        $attributes->setPosition(30);
-        $attributes->setComponent('products/components/attributes@suluproduct');
-        $attributes->setDisplay(['edit']);
-        $attributes->setResetStore(false);
-
-        return [
-            $details,
-            $variants,
-            $attributes,
-        ];
+        return array_merge(
+            $navigationItems,
+            [
+                $variants,
+            ]
+        );
     }
 }