+++ Resources/public/js/components/products/components/attributes/main.js
@@ -1,5 +1,5 @@
 /*
- * This file is part of the Sulu CMS.
+ * This file is part of Sulu.
  *
  * (c) MASSIVE ART WebServices GmbH
  *
@@ -9,85 +9,170 @@
 
 define([
     'config',
-    'text!suluproduct/components/products/components/attributes/overlay-content.html'
-], function(Config, OverlayTpl) {
+    'text!suluproduct/components/products/components/attributes/overlay-content.html',
+    'services/product-type-manager',
+    'suluproduct/collections/attributes',
+    'suluproduct/models/variantAttribute'
+], function(Config, OverlayTpl, ProductTypeManager, Attributes, VariantAttribute) {
+
     'use strict';
 
-    var formSelector = '#product-attributes-form',
-        productAttributesInstanceName = 'product-attribute-list-toolbar',
-        datagridInstanceName = 'product-attribute-datagrid',
-        overlayInstanceName = 'product-attribute-overlay',
-        selectInstanceName = 'product-attribute-select',
-        typeText = 'product.attribute.type.text',
-        attributeId = null,
+    var selectedAttributeId = null,
+        selectedAttributeName = null,
         actions = {
             ADD: 1,
             DELETE: 2,
             UPDATE: 3
         },
 
-        /**
-         * bind custom events
-         */
-        bindCustomEvents = function() {
-            this.sandbox.on('sulu.toolbar.delete', function() {
-                this.sandbox.emit('sulu.product.delete', this.options.data.id);
-            }.bind(this));
+        constants = {
+            typeText: 'product.attribute.type.text',
+            datagridInstanceName: 'product-attribute-datagrid',
+            variantAttributesDatagridInstanceName: 'variant-attribute-datagrid',
+            attributesToolbarInstanceName: 'product-attribute-list-toolbar',
+            variantAttributesToolbarInstanceName: 'variant-attribute-list-toolbar',
+            overlayInstanceName: 'product-attribute-overlay',
+            selectInstanceName: 'product-attribute-select'
+        },
 
-            this.sandbox.on('product.state.change', function(status) {
-                if (!this.options.data ||
-                    !this.options.data.attributes.status ||
-                    this.options.data.attributes.status.id !== status.id
-                ) {
-                    this.status = status;
-                    this.options.data.attributes.status = this.status;
-                    setHeaderBar.call(this, false);
-                }
-            }, this);
 
-            this.sandbox.on('sulu.toolbar.save', function() {
-                this.sendData = {};
-                this.sendData.status = this.status;
-                this.sendData.id = this.options.data.id;
-                save.call(this);
-            }, this);
-
-            this.sandbox.on('sulu.products.saved', function(data) {
-
-                var attributes = data.attributes;
-
-                // Select action
-                if (data.action === actions.ADD) {
-                    // ADD RECORD IN DATAGRID
-                    var attribute = _.findWhere(attributes, {'attributeId': data.attributeIdAdded});
-                    this.sandbox.emit('husky.datagrid.' + datagridInstanceName + '.record.add', attribute);
-                } else if (data.action === actions.DELETE) {
-                    // DELETE RECORDs IN DATAGRID
-                    $.each(data.attributeIdsDeleted, function(key, id) {
-                        this.sandbox.emit('husky.datagrid.' + datagridInstanceName + '.record.remove', id);
-                    }.bind(this));
-                } else if (data.action === actions.UPDATE) {
-                    // UPDATE DATAGRID WITH RECEIVED RECORDS
-                    this.sandbox.emit('husky.datagrid.' + datagridInstanceName + '.records.set', attributes);
-                }
-
-                setHeaderBar.call(this, true);
-                //this.options.data = data;
-                this.options.data.attributes.status = this.status;
-            }, this);
-
-            // enable toolbar items
-            this.sandbox.on('husky.datagrid.' + datagridInstanceName + '.number.selections', function(number) {
-                var postfix = number > 0 ? 'enable' : 'disable';
-                this.sandbox.emit(
-                    'husky.toolbar.' + productAttributesInstanceName + '.item.' + postfix,
-                    'delete',
-                    false)
-            }, this);
+        selectors = {
+            variantAttributesContainer: '#js-variant-attributes-container'
         },
 
         /**
-         * @param {Boolean} saved defines if saved state should be shown
+         * Fetches all attributes from database and parses them for beeing displayed in a select.
+         *
+         * @returns {Object} Deferred promise
+         */
+        fetchAttributesForSelect = function() {
+            var deferred = $.Deferred();
+            var attributes = new Attributes({locale: this.options.locale});
+            attributes.fetch({
+                success: function(data) {
+                    var selectAttributes = [];
+                    $.each(data.toJSON(), function(key, value) {
+                        var attribute = {
+                            'id': value.id,
+                            'name': value.name
+                        };
+
+                        // At this time we support only text type attributes.
+                        if (value.type.name === constants.typeText) {
+                            selectAttributes.push(attribute);
+                        }
+                    }.bind(this));
+
+                    deferred.resolve(selectAttributes);
+                }.bind(this),
+                error: function() {
+                    console.error('Error retrieving attributes from server');
+                    deferred.fail();
+                }
+            });
+
+            return deferred.promise();
+        },
+
+        /**
+         * Bind custom events.
+         */
+        bindCustomEvents = function() {
+            this.sandbox.on('sulu.toolbar.delete', onDeleteClicked.bind(this));
+
+            this.sandbox.on('product.state.change', onStatusChanged.bind(this));
+
+            this.sandbox.on('sulu.toolbar.save', onSaveClicked.bind(this));
+
+            this.sandbox.on('sulu.products.saved', onProductSaved.bind(this));
+
+            this.sandbox.on(
+                'husky.datagrid.' + constants.datagridInstanceName + '.number.selections',
+                onProductAttributeSelection.bind(this, constants.attributesToolbarInstanceName)
+            );
+
+            this.sandbox.on(
+                'husky.datagrid.' + constants.variantAttributesDatagridInstanceName + '.number.selections',
+                onProductAttributeSelection.bind(this, constants.variantAttributesToolbarInstanceName)
+            );
+        },
+
+        /**
+         * Called when delete button is clicked.
+         */
+        onDeleteClicked = function() {
+            this.sandbox.emit('sulu.product.delete', this.options.data.id);
+        },
+
+        /**
+         * Enables or disables toolbar based on number of items that were selected.
+         *
+         * @param {Number} number
+         * @param {String} toolBarInstanceName
+         */
+        onProductAttributeSelection = function(toolBarInstanceName, number) {
+            var postfix = number > 0 ? 'enable' : 'disable';
+            this.sandbox.emit(
+                'husky.toolbar.' + toolBarInstanceName + '.item.' + postfix,
+                'delete',
+                false)
+        },
+
+        /**
+         * Callback, when product has been saved.
+         *
+         * @param {Object} data
+         */
+        onProductSaved = function(data) {
+            var attributes = data.attributes;
+
+            // Select action.
+            if (data.action === actions.ADD) {
+                // Add records in datagrid.
+                var attribute = _.findWhere(attributes, {'attributeId': data.attributeIdAdded});
+                this.sandbox.emit('husky.datagrid.' + constants.datagridInstanceName + '.record.add', attribute);
+            } else if (data.action === actions.DELETE) {
+                // Delete records in datagrid.
+                $.each(data.attributeIdsDeleted, function(key, id) {
+                    this.sandbox.emit('husky.datagrid.' + constants.datagridInstanceName + '.record.remove', id);
+                }.bind(this));
+            } else if (data.action === actions.UPDATE) {
+                // Update datagrid with received records.
+                this.sandbox.emit('husky.datagrid.' + constants.datagridInstanceName + '.records.set', attributes);
+            }
+
+            setHeaderBar.call(this, true);
+            this.options.data.attributes.status = this.status;
+        },
+
+        /**
+         * Called when save button was clicked.
+         */
+        onSaveClicked = function() {
+            this.sendData = {};
+            this.sendData.status = this.status;
+            this.sendData.id = this.options.data.id;
+            save.call(this);
+        },
+
+        /**
+         * Called when product status has changed.
+         *
+         * @param {Object} status
+         */
+        onStatusChanged = function(status) {
+            if (!this.options.data
+                || !this.options.data.attributes.status
+                || this.options.data.attributes.status.id !== status.id
+            ) {
+                this.status = status;
+                this.options.data.attributes.status = this.status;
+                setHeaderBar.call(this, false);
+            }
+        },
+
+        /**
+         * @param {Boolean} saved Defines if saved state should be shown.
          */
         setHeaderBar = function(saved) {
             if (saved !== this.saved) {
@@ -101,14 +186,17 @@ define([
         },
 
         /**
-         * Create overlay content for add attribute overlay
+         * Create overlay content for add attribute overlay.
+         *
+         * @param {Boolean} shouldHideValueField
          */
-        createAddOverlayContent = function() {
-            attributeId = null;
+        createAddOverlayContent = function(shouldHideValueField) {
+            selectedAttributeId = null;
 
-            // create container for overlay
+            // Create container for overlay.
             var $overlayContent = this.sandbox.dom.createElement(this.sandbox.util.template(OverlayTpl, {
-                translate: this.sandbox.translate
+                translate: this.sandbox.translate,
+                shouldHideValueField: shouldHideValueField
             }));
             this.sandbox.dom.append(this.$el, $overlayContent);
 
@@ -116,98 +204,21 @@ define([
         },
 
         /**
-         * Create the overlay
+         * Creates the overlay.
          */
-        createAddOverlay = function() {
-            // call JSON to get the attributes from the server then create the overlay after it's done
-            var attributesUrl = 'api/attributes?locale=' + this.options.locale;
-            var ajaxRequest = $.getJSON(attributesUrl, function(data) {
-                this.attributeTypes = [];
-
-                $.each(data._embedded.attributes, function(key, value) {
-                    var newAttribute = {
-                        'id': value.id,
-                        'name': value.name
-                    };
-
-                    // at this time we support only text type attributes
-                    if (value.type.name === typeText) {
-                        this.attributeTypes.push(newAttribute);
-                    }
-                }.bind(this));
-            }.bind(this));
-
-            ajaxRequest.done(function() {
-                // create container for overlay
-                var $overlay = this.sandbox.dom.createElement('<div>');
-                this.sandbox.dom.append(this.$el, $overlay);
-
-                // create content
-                this.sandbox.start([
-                    {
-                        name: 'overlay@husky',
-                        options: {
-                            el: $overlay,
-                            supportKeyInput: false,
-                            title: this.sandbox.translate('product.attribute.overlay.title'),
-                            skin: 'normal',
-                            openOnStart: true,
-                            removeOnClose: true,
-                            instanceName: overlayInstanceName,
-                            data: createAddOverlayContent.call(this),
-                            okCallback: overlayOkClicked.bind(this)
-                        }
-                    }
-                ]);
-
-            }.bind(this));
-
-            ajaxRequest.fail(function() {
-                console.log('Error retrieving attributes from server');
-            }.bind(this));
-
-            ajaxRequest.complete(function() {
-
-                var preSelectedElement = [];
-
-                // set pre selected element in checkbox
-                if (this.attributeTypes.length > 0 &&
-                    typeof(this.attributeTypes[0]) === "object" &&
-                    typeof(this.attributeTypes[0].name) === "string"
-                ) {
-                    attributeId = this.attributeTypes[0].id;
-                    preSelectedElement.push(this.attributeTypes[0].name);
-                }
-
-                // create dropbox in overlay
-                var selectOptions = {
-                    el: '#selectBox',
-                    instanceName: selectInstanceName,
-                    multipleSelect: false,
-                    defaultLabel: this.sandbox.translate('product.attribute.overlay.defaultlabel'),
-                    preSelectedElements: preSelectedElement,
-                    valueName: 'name',
-                    isNative: true,
-                    data: this.attributeTypes
-                };
-
-                this.sandbox.start([
-                    // start select
-                    {
-                        name: 'select@husky',
-                        options: selectOptions
-                    }
-                ]);
-
-                // define select event for dropbox
-                this.sandbox.on('husky.select.' + selectInstanceName + '.selected.item', function(item) {
-                    attributeId = parseInt(item);
-                });
+        onAddAttributeClicked = function() {
+            fetchAttributesForSelect.call(this).done(function(selectData) {
+                createAddOverlay.call(
+                    this,
+                    createAddOverlayContent.call(this, false),
+                    onAddAttributeOkClicked.bind(this)
+                );
+                startAttributesSelect.call(this, selectData);
             }.bind(this));
         },
 
         /**
-         * save product attributes
+         * Save product attributes.
          */
         save = function() {
             this.saved = false;
@@ -215,11 +226,11 @@ define([
         },
 
         /**
-         * called when OK on overlay was clicked
+         * Called when attribute and value have been selected in overlay.
          */
-        overlayOkClicked = function() {
-            // exit if no attribute is selected in overlay
-            if (!attributeId) {
+        onAddAttributeOkClicked = function() {
+            // Exit if no attribute is selected in overlay.
+            if (!selectedAttributeId) {
                 return;
             }
 
@@ -228,7 +239,7 @@ define([
 
             var attributes = this.options.data.attributes.attributes;
 
-            var result = _.findWhere(attributes, {'attributeId': attributeId});
+            var result = _.findWhere(attributes, {'attributeId': selectedAttributeId});
 
             if (result) {
                 result.attributeValueName = attributeValueName;
@@ -236,7 +247,7 @@ define([
                 this.sendData.action = actions.UPDATE;
             } else {
                 var newAttribute = {
-                    'attributeId': attributeId,
+                    'attributeId': selectedAttributeId,
                     'attributeValueName': attributeValueName,
                     'attributeValueLocale': this.options.locale
                 };
@@ -244,7 +255,7 @@ define([
                 this.sendData.action = actions.ADD;
             }
 
-            this.sendData.attributeIdAdded = attributeId;
+            this.sendData.attributeIdAdded = selectedAttributeId;
             this.sendData.attributes = attributes;
             this.sendData.status = this.status;
             this.sendData.id = this.options.data.attributes.id;
@@ -253,16 +264,129 @@ define([
         },
 
         /**
-         * delete action function from toolbar
+         * Called when attribute has been selected in add variant attribute overlay.
+         */
+        onAddVariantAttributeOkClicked = function() {
+            // Exit if no attribute is selected in overlay.
+            if (!selectedAttributeId || !selectedAttributeName) {
+                return;
+            }
+
+            var variantAttribute = new VariantAttribute({
+                productId: this.options.data.id,
+                attributeId: selectedAttributeId
+            });
+            variantAttribute.save(
+                {},
+                {
+                    success: onAddVariantAttributeSuccess.bind(this),
+                    error: onAddVariantAttributeError.bind(this)
+                }
+            );
+        },
+
+        /**
+         * Called when delete variant attribute was clicked.
+         */
+        onDeleteVariantAttributeClicked = function() {
+            var selectedIds = [];
+            this.sandbox.emit(
+                'husky.datagrid.' + constants.variantAttributesDatagridInstanceName + '.items.get-selected',
+                function(ids) {
+                    selectedIds = ids;
+                }
+            );
+
+            var numberOfDeletions = selectedIds.length;
+            if (numberOfDeletions < 1) {
+                return;
+            }
+
+            var successCounter = 0;
+            this.sandbox.util.foreach(selectedIds, function(value) {
+                var variantAttribute = new VariantAttribute({
+                    productId: this.options.data.id,
+                    id: value
+                });
+
+                variantAttribute.destroy({
+                    success: function() {
+                        successCounter++;
+                        onDeleteVariantAttributeSuccess.call(this, value);
+                        checkDeleteSuccess.call(this, successCounter, numberOfDeletions);
+                    }.bind(this),
+                    error: onDeleteVariantAttributeError.bind(this)
+                });
+            }.bind(this));
+        },
+
+        /**
+         * Called when post variant attribute was successful.
+         */
+        onAddVariantAttributeSuccess = function() {
+            // TODO: translation
+            this.sandbox.emit('sulu.labels.success.show', 'labels.success.save-desc', 'labels.success');
+            this.sandbox.emit(
+                'husky.datagrid.' + constants.variantAttributesDatagridInstanceName + '.record.add',
+                {
+                    'id': selectedAttributeId,
+                    'name': selectedAttributeName
+                }
+            );
+        },
+
+        /**
+         * Called when post variant attribute has failed.
+         */
+        onAddVariantAttributeError = function() {
+            this.sandbox.emit('sulu.labels.error.show', 'sulu_product.labels.save-error', 'labels.error');
+        },
+
+        /**
+         * Removes deleted attribute from variants datagrid.
+         *
+         * @param {Number} id
+         */
+        onDeleteVariantAttributeSuccess = function(id) {
+            // Remove record from datagrid.
+            this.sandbox.emit(
+                'husky.datagrid.' + constants.variantAttributesDatagridInstanceName + '.record.remove',
+                id
+            );
+        },
+
+        /**
+         * Shows delete success label when current number of deletion matches expected count.
+         *
+         * @param {Number} currentCount
+         * @param {Number} expectedCount
+         */
+        checkDeleteSuccess = function(currentCount, expectedCount) {
+            if (currentCount !== expectedCount) {
+                return;
+            }
+
+            this.sandbox.emit('sulu.labels.success.show', 'labels.success.delete-desc', 'labels.success');
+        },
+
+        /**
+         * Called when delete request is not successful.
+         */
+        onDeleteVariantAttributeError = function() {
+            this.sandbox.emit('sulu.labels.error.show', 'sulu_product.labels.delete-error', 'labels.error');
+        },
+
+        /**
+         * Delete action function from toolbar.
          */
         attributeDelete = function() {
-            this.sandbox.emit('husky.datagrid.' + datagridInstanceName + '.items.get-selected', function(ids) {
+            this.sandbox.emit('husky.datagrid.' + constants.datagridInstanceName + '.items.get-selected', function(ids) {
 
                 var attributes = this.options.data.attributes.attributes;
                 this.sendData = {};
                 var attributeIdsDeleted = [];
 
-                _.each(ids, function(value, key, list) {
+                _.each(ids, function(value) {
                     var result = _.findWhere(attributes, {'attributeId': value});
                     attributes = _.without(attributes, result);
                     attributeIdsDeleted.push(value);
@@ -286,8 +410,7 @@ define([
          * @param {String} locale
          */
         onBadgeAttributeName = function(item, badge, locale) {
-            if (
-                item.attributeLocale
+            if (item.attributeLocale
                 && item.attributeLocale == item.fallbackLocale
                 && item.attributeLocale != locale
             ) {
@@ -307,8 +430,7 @@ define([
          * @param {String} locale
          */
         onBadgeAttributeValueName = function(item, badge, locale) {
-            if (
-                item.attributeValueLocale
+            if (item.attributeValueLocale
                 && item.attributeValueLocale == item.fallbackLocale
                 && item.attributeValueLocale != locale
             ) {
@@ -321,12 +443,12 @@ define([
         },
 
         /**
-         * calls basic form components
+         * Calls basic form components.
          */
         startFormComponents = function() {
             var datagridOptions = {
                 el: '#product-attribute-list',
-                instanceName: datagridInstanceName,
+                instanceName: constants.datagridInstanceName,
                 idKey: 'attributeId',
                 resultKey: 'attributes',
                 matchings: [
@@ -362,25 +484,22 @@ define([
             };
 
             this.sandbox.start([
-                // start datagrid
+                // Start datagrid.
                 {
                     name: 'datagrid@husky',
                     options: datagridOptions
-                }
-            ]);
-
-            this.sandbox.start([
+                },
                 {
                     name: 'toolbar@husky',
                     options: {
-                        el: '#product-attribute-selection-toolbar',
-                        instanceName: productAttributesInstanceName,
+                        el: '#product-attribute-toolbar',
+                        instanceName: constants.attributesToolbarInstanceName,
                         small: false,
                         buttons: [
                             {
                                 id: 'add',
                                 icon: 'plus-circle',
-                                callback: createAddOverlay.bind(this)
+                                callback: onAddAttributeClicked.bind(this)
                             },
                             {
                                 id: 'delete',
@@ -395,10 +514,180 @@ define([
         },
 
         /**
-         * initialize form components
+         * Starts and displays overlay component for adding variant attributes.
          */
-        initForm = function() {
-            this.sandbox.start(formSelector);
+        createAddOverlay = function(overlayContent, okCallback) {
+            // Create container for overlay.
+            var $overlay = this.sandbox.dom.createElement('<div>');
+            this.sandbox.dom.append(this.$el, $overlay);
+
+            // Create content.
+            this.sandbox.start([
+                {
+                    name: 'overlay@husky',
+                    options: {
+                        el: $overlay,
+                        supportKeyInput: false,
+                        title: this.sandbox.translate('product.attribute.overlay.title'),
+                        skin: 'normal',
+                        openOnStart: true,
+                        removeOnClose: true,
+                        instanceName: constants.overlayInstanceName,
+                        data: overlayContent,
+                        okCallback: okCallback
+                    }
+                }
+            ]);
+        },
+
+        /**
+         * Starts the select component with given data.
+         *
+         * @param {Array} selectData
+         */
+        startAttributesSelect = function(selectData) {
+            var preSelectedElement = [];
+            var defaultLabel = this.sandbox.translate('product.attribute.overlay.defaultlabel');
+
+            // Preselect first element.
+            if (selectData.length > 0 &&
+                typeof(selectData[0]) === "object" &&
+                typeof(selectData[0].name) === "string"
+            ) {
+                selectedAttributeId = selectData[0].id;
+                preSelectedElement.push(selectData[0].name);
+            } else {
+                defaultLabel = this.sandbox.translate('sulu_product.attribute.overlay.no-attributes')
+            }
+
+            // Create select box in overlay.
+            var selectOptions = {
+                el: '#selectBox',
+                instanceName: constants.selectInstanceName,
+                multipleSelect: false,
+                defaultLabel: defaultLabel,
+                preSelectedElements: preSelectedElement,
+                valueName: 'name',
+                isNative: true,
+                data: selectData
+            };
+
+            this.sandbox.start([
+                {
+                    name: 'select@husky',
+                    options: selectOptions
+                }
+            ]);
+
+            // Define callback when attribute is selected.
+            this.sandbox.on('husky.select.' + constants.selectInstanceName + '.selected.item', function(item, name) {
+                selectedAttributeId = parseInt(item);
+                selectedAttributeName = name;
+            });
+        },
+
+        /**
+         * Gets called when add button for adding an attribute was clicked.
+         */
+        onAddVariantAttributeClicked = function() {
+            fetchAttributesForSelect.call(this).done(function(selectData) {
+                // Get all Attributes that already have been added.
+                var alreadyAddedAttributeIds = [];
+                this.sandbox.emit(
+                    'husky.datagrid.' + constants.variantAttributesDatagridInstanceName + '.records.get',
+                    function(data) {
+                        alreadyAddedAttributeIds = data.map(function(attribute) {
+                            return attribute.id;
+                        });
+                    }.bind(this)
+                );
+                // Do not add attributes that already are added.
+                var filteredSelectData = [];
+                this.sandbox.util.foreach(selectData, function(data) {
+                    if (alreadyAddedAttributeIds.indexOf(data.id) === -1) {
+                        filteredSelectData.push(data);
+                    }
+                });
+
+                createAddOverlay.call(
+                    this,
+                    createAddOverlayContent.call(this, true),
+                    onAddVariantAttributeOkClicked.bind(this)
+                );
+                startAttributesSelect.call(this, filteredSelectData);
+            }.bind(this));
+        },
+
+        /**
+         * Starts toolbar and datagrid for managing variant attributes.
+         */
+        startVariantAttributeFormComponents = function() {
+            var datagridOptions = {
+                el: '#js-variant-attribute-list',
+                instanceName: constants.variantAttributesDatagridInstanceName,
+                resultKey: 'variantAttributes',
+                matchings: '/admin/api/product-variant-attributes/fields?locale=' + this.options.locale,
+                url: '/admin/api/products/' + this.options.data.id + '/variant-attributes?locale=' + this.options.locale
+            };
+
+            this.sandbox.start([
+                // Start datagrid.
+                {
+                    name: 'datagrid@husky',
+                    options: datagridOptions
+                },
+
+            ]);
+
+            // Only show toolbar if no variants have been added yet.
+            if (this.options.data.attributes.numberOfVariants === 0) {
+                this.sandbox.start([
+                    {
+                        name: 'toolbar@husky',
+                        options: {
+                            el: '#js-variant-attribute-toolbar',
+                            instanceName: constants.variantAttributesToolbarInstanceName,
+                            small: false,
+                            buttons: [
+                                {
+                                    id: 'add',
+                                    icon: 'plus-circle',
+                                    callback: onAddVariantAttributeClicked.bind(this)
+                                },
+                                {
+                                    id: 'delete',
+                                    icon: 'trash-o',
+                                    disabled: true,
+                                    callback: onDeleteVariantAttributeClicked.bind(this)
+                                }
+                            ]
+                        }
+                    }
+                ]);
+            }
+        },
+
+        /**
+         * Initialize variant attributes components.
+         */
+        initVariantAttributesForm = function() {
+            var productType = this.options.data.attributes.type.id;
+
+            // If current product is a product with variants we also show product variants table.
+            if (productType !== ProductTypeManager.types.PRODUCT_WITH_VARIANTS) {
+                return;
+            }
+
+            // Show variant attributes container.
+            $(selectors.variantAttributesContainer).removeClass('is-hidden');
+
+            startVariantAttributeFormComponents.call(this);
+        },
+
+        /**
+         * Initialize product attributes components.
+         */
+        initProductAttributesForm = function() {
             startFormComponents.call(this);
         };
 
@@ -407,22 +696,40 @@ define([
 
         templates: ['/admin/product/template/product/attributes'],
 
-        render: function() {
-            this.sandbox.dom.html(this.$el, this.renderTemplate('/admin/product/template/product/attributes'));
-            initForm.call(this);
-        },
-
+        /**
+         * Constructor of component.
+         */
         initialize: function() {
             bindCustomEvents.call(this);
+
+            // Set correct status.
+            this.status = Config.get('product.status.inactive');
             if (!!this.options.data) {
                 this.status = this.options.data.attributes.status;
-            } else {
-                this.status = Config.get('product.status.inactive');
             }
-            // reset status if it has been changed before and has not been saved
+            // Reset status if it has been changed before and has not been saved.
             this.sandbox.emit('product.state.change', this.status);
+
             this.render();
             setHeaderBar.call(this, true);
+        },
+
+        /**
+         * Renders component.
+         */
+        render: function() {
+            this.sandbox.dom.html(
+                this.$el,
+                this.renderTemplate(
+                    '/admin/product/template/product/attributes',
+                    {
+                        'translate': this.sandbox.translate
+                    }
+                )
+            );
+
+            initProductAttributesForm.call(this);
+            initVariantAttributesForm.call(this);
         }
     };
 });